+++ ConsoleAppAids3/Properties/AssemblyInfo.cs
@@ -108,6 +108,6 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion ( "8.1.560.0" )]
-[assembly: AssemblyFileVersion ( "8.1.570.0" )]
-[assembly: AssemblyInformationalVersion ( "8.1.570" )]
+[assembly: AssemblyVersion ( "9.0.0.0" )]
+[assembly: AssemblyFileVersion ( "9.0.566.0" )]
+[assembly: AssemblyInformationalVersion ( "9.0.566" )]