+++ EOJTEST/Properties/AssemblyInfo.cs
@@ -98,4 +98,6 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion ( "8.1.530.0" )]
+[assembly: AssemblyVersion ( "9.0.0.0" )]
+[assembly: AssemblyFileVersion ( "9.0.534.0" )]
+[assembly: AssemblyInformationalVersion ( "9.0.534" )]