+++ TestStand/Properties/AssemblyInfo.cs
@@ -46,6 +46,9 @@
 
     2021/06/06 8.0.535 DAG Build against the most recent WizardWrx .NET API
                            library constellation.
+
+    2023/09/09 9.0.579 DAG Update the copyright year, coinciding with the target
+                           framework upgrade.
     ============================================================================
 */
 
@@ -61,7 +64,7 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration ( "" )]
 [assembly: AssemblyCompany ( "David A. Gray" )]
 [assembly: AssemblyProduct ( "WizardWrx .NET Utility Libraries" )]
-[assembly: AssemblyCopyright ( "Copyright 2014-2022, David A. Gray" )]
+[assembly: AssemblyCopyright ( "Copyright 2014-2023, David A. Gray" )]
 [assembly: AssemblyTrademark ( "This library is distributed under a three-clause BSD license." )]
 [assembly: AssemblyCulture ( "" )]
 
@@ -84,5 +87,5 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion ( "9.0.0.0" )]
-[assembly: AssemblyFileVersion ( "9.0.578.0" )]
-[assembly: AssemblyInformationalVersion ( "9.0.578" )]
+[assembly: AssemblyFileVersion ( "9.0.579.0" )]
+[assembly: AssemblyInformationalVersion ( "9.0.579" )]