+++ ChangeLog.md
@@ -4,6 +4,11 @@ This file is a running history of fixes and improvements from version 7.0
 onwards. Changes are documented for the newest version first. Within each
 version, classes are listed alphabetically.
 
+# AssemblyVersion 9.0.0 Updated 12/16/2023, AssemblyFileVersion 9.0.569
+
+Unwind screen scroll suppression in `WaitForCarbonUnit` that was implemented in
+version 6.3, as it has proven troublesome in shell scripts.
+
 # AssemblyVersion 9.0.0 Updated 09/09/2023, AssemblyFileVersion 9.0.567
 
 To maintain compatibility with the WizardWrx .NET API upon which this assembly